% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/als_implicit.R
\docType{data}
\name{ALS_implicit}
\alias{ALS_implicit}
\title{ALS implicit}
\format{\code{\link{R6Class}} object.}
\usage{
ALS_implicit
}
\description{
Creates ALS implicit model.
See (Hu, Koren, Volinsky)'2008 paper \url{http://yifanhu.net/PUB/cf.pdf} for details.
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
asl_impl = ALS_implicit$new(rank = 10L, lambda = 0, init_stdv = ifelse(lambda == 0, 0.01, 1 / sqrt(2 * lambda)))
asl_impl$fit(x, n_iter = 5L, n_cores = 1, trace = FALSE, ...)
asl_impl$fit_transform(x, n_iter = 5L, n_cores = 1, trace = FALSE, ...)
asl_impl$components
}
}

\section{Methods}{

\describe{
  \item{\code{$new(rank = 10L, lambda = 0, init_stdv = ifelse(lambda == 0, 0.01, 1 / sqrt(2 * lambda)))}}{
    create ALS_implicit model with \code{rank} latent factors}
  \item{\code{$fit(x, n_iter = 5L, n_cores = 1, trace = FALSE, ...)}}{
    fit model to an input user-item \bold{confidence!} matrix. (preferably in "dgCMatrix" format)}.
    \code{x} should be a confidence matrix which corresponds to \code{1 + alpha * r_ui} in original paper.
    Usually \code{r_ui} cosrresponds to the number of interactions of user \code{u} and item \code{i}.
  \item{\code{$fit_transform(x, n_iter = 5L, n_cores = 1, trace = FALSE, ...)}}{Explicitly returns factor matrix for
    users of size \code{n_users * rank}. See description for \code{fit} above.}.
  \item{\code{$components}}{item factors matrix of size \code{rank * n_items}}.
}
}

\section{Arguments}{

\describe{
 \item{asl_impl}{A \code{ALS_implicit} model.}
 \item{x}{An input user-item \bold{confidence} matrix.}
 \item{rank}{\code{integer} desired number of latent factors}
 \item{lambda}{\code{numeric} regularization parameter}
 \item{init_stdv}{\code{numeric} std dev for initialization of the factor matrices}
 \item{trace}{\code{logical} whether to calculate loss proxy. By "proxy" we mean that we will calculate
 loss only for user-item pairs with observer interactions}
 \item{n_cores}{\code{n_cores} number of cores to use in factorization. Corresponds to
   \code{mc.cores} in \code{parallel::mclapply}. Ignored for Windows OS (with warning).}
 \item{...}{Arguments useful for \code{fit(), fit_transform()} -
 these arguments will be passed to \link{parallel::mclapply} function which is used for parallelization
 of computations.}
}
}
\keyword{datasets}

