% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{detect_number_omp_threads}
\alias{detect_number_omp_threads}
\title{Detects number of OpenMP threads in the system}
\usage{
detect_number_omp_threads()
}
\description{
Detects number of OpenMP threads in the system respecting environment
variables such as \code{OMP_NUM_THREADS} and \code{OMP_THREAD_LIMIT}
}
