% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mathematical-functions}
\alias{mathematical-functions}
\alias{sqrt,RsparseMatrix-method}
\alias{abs,RsparseMatrix-method}
\alias{log1p,RsparseMatrix-method}
\alias{cos,RsparseMatrix-method}
\alias{tan,RsparseMatrix-method}
\alias{tanh,RsparseMatrix-method}
\alias{tanpi,RsparseMatrix-method}
\alias{sinh,RsparseMatrix-method}
\alias{atanh,RsparseMatrix-method}
\alias{expm1,RsparseMatrix-method}
\alias{sign,RsparseMatrix-method}
\alias{ceiling,RsparseMatrix-method}
\alias{floor,RsparseMatrix-method}
\alias{trunc,RsparseMatrix-method}
\alias{round,RsparseMatrix-method}
\alias{signif,RsparseMatrix-method}
\title{Mathematical functions for CSR matrices}
\usage{
\S4method{sqrt}{RsparseMatrix}(x)

\S4method{abs}{RsparseMatrix}(x)

\S4method{log1p}{RsparseMatrix}(x)

\S4method{cos}{RsparseMatrix}(x)

\S4method{tan}{RsparseMatrix}(x)

\S4method{tanh}{RsparseMatrix}(x)

\S4method{tanpi}{RsparseMatrix}(x)

\S4method{sinh}{RsparseMatrix}(x)

\S4method{atanh}{RsparseMatrix}(x)

\S4method{expm1}{RsparseMatrix}(x)

\S4method{sign}{RsparseMatrix}(x)

\S4method{ceiling}{RsparseMatrix}(x)

\S4method{floor}{RsparseMatrix}(x)

\S4method{trunc}{RsparseMatrix}(x)

\S4method{round}{RsparseMatrix}(x, digits = 0)

\S4method{signif}{RsparseMatrix}(x, digits = 6)
}
\arguments{
\item{x}{A CSR matrix.}

\item{digits}{See \link{round} and \link{signif}. If passing more than one value,
will call the corresponding function from the `Matrix` package, which implies first
converting `x` to CSC format.}
}
\value{
A CSR matrix in `dgRMatrix` format.
}
\description{
Implements some mathematical functions for CSR matrices
(a.k.a. "RsparseMatrix") without converting them to CSC matrices in the process.

These functions reduce to applying the same function over the non-zero elements only,
and as such do not benefit from any storage format conversion as done implicitly
in the `Matrix` package.
}
\examples{
library(Matrix)
library(rsparse)
set.seed(1)
X = as.csr.matrix(rsparsematrix(4, 3, .4))
abs(X)
sqrt(X^2)
}
