% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{matmult}
\alias{matmult}
\alias{\%*\%,dgRMatrix,matrix-method}
\alias{tcrossprod,dgRMatrix,matrix-method}
\alias{\%*\%,matrix,dgCMatrix-method}
\alias{crossprod,matrix,dgCMatrix-method}
\title{Multithreaded Sparse-Dense Matrix Multiplication}
\usage{
\S4method{\%*\%}{dgRMatrix,matrix}(x, y)

\S4method{tcrossprod}{dgRMatrix,matrix}(x, y)

\S4method{\%*\%}{matrix,dgCMatrix}(x, y)

\S4method{crossprod}{matrix,dgCMatrix}(x, y)
}
\arguments{
\item{x, y}{dense \code{matrix} and sparse
\code{Matrix::RsparseMatrix} / \code{Matrix::CsparseMatrix} matrices.}
}
\value{
A dense \code{matrix}
}
\description{
Multithreaded \code{\%*\%}, \code{crossprod}, \code{tcrossprod}
for sparse-dense matrix multiplication
}
\details{
Accelerates sparse-dense matrix multiplications using openmp. Applicable to the following pairs:
(\code{dgRMatrix}, \code{matrix}), (\code{matrix}, \code{dgRMatrix}),
(\code{dgCMatrix}, \code{matrix}), (\code{matrix}, \code{dgCMatrix}) combinations
}
