% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{movielens100k}
\alias{movielens100k}
\title{MovieLens 100K Dataset}
\format{A sparse column-compressed matrix (\code{Matrix::dgCMatrix}) with 943 rows and 1682 columns.
\enumerate{
  \item rows are users
  \item columns are movies
  \item values are ratings
}}
\source{
\url{https://en.wikipedia.org/wiki/MovieLens#Datasets}
}
\usage{
data("movielens100k")
}
\description{
This data set consists of:
\enumerate{
  \item 100,000 ratings (1-5) from 943 users on 1682 movies.
  \item Each user has rated at least 20 movies.
}
MovieLens data sets were collected by the GroupLens Research Project at the University of Minnesota.
}
\keyword{datasets}
