% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftALS.R
\name{soft_impute}
\alias{soft_impute}
\alias{soft_svd}
\title{SoftImpute/SoftSVD matrix factorization}
\usage{
soft_impute(x, rank = 10L, lambda = 0, n_iter = 100L,
  convergence_tol = 0.001, init = NULL, final_svd = TRUE)

soft_svd(x, rank = 10L, lambda = 0, n_iter = 100L,
  convergence_tol = 0.001, init = NULL, final_svd = TRUE)
}
\arguments{
\item{x}{sparse matrix. Both CSR \code{dgRMatrix} and CSC \code{dgCMatrix} are supported.
in case of CSR matrix we suggest to load \url{https://github.com/dselivanov/MatrixCSR} package
which provides multithreaded CSR*dense matrix products (if OpenMP is supported on your platform).
On many-cores machines this reduces fitting time significantly.}

\item{rank}{maximum rank of the low-rank solution.}

\item{lambda}{regularization parameter for nuclear norm}

\item{n_iter}{maximum number of iterations of the algorithms}

\item{convergence_tol}{convergence tolerance.
Internally we keep track relative change of frobenious norm of two consequent iterations.}

\item{init}{\link{svd} like object with \code{u, v, d} components to initialize algorithm.
Algorithm benefit from warm starts. \code{init} could be rank up \code{rank} of the maximum allowed rank.
If \code{init} has rank less than max rank it will be padded automatically.}

\item{final_svd}{\code{logical} whether need to make final preprocessing with SVD.
This is not necessary but cleans up rank nicely - hithly recommnded to leave it \code{TRUE}.}
}
\value{
\link{svd}-like object - \code{list} with \code{u, v, d}
components - left, right singular vectors and singular vectors.
}
\description{
Fit SoftImpute/SoftSVD via fast alternating least squares. Based on the
paper by Trevor Hastie, Rahul Mazumder, Jason D. Lee, Reza Zadeh
by "Matrix Completion and Low-Rank SVD via Fast Alternating Least Squares" -
\url{https://arxiv.org/pdf/1410.2596.pdf}
}
