% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_FactorizationMachine.R
\docType{data}
\name{FactorizationMachine}
\alias{FactorizationMachine}
\title{Creates FactorizationMachine model.}
\format{\code{\link{R6Class}} object.}
\usage{
FactorizationMachine
}
\description{
Creates second order Factorization Machines model
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
fm = FM$new(learning_rate = 0.2, rank = 8, lambda_w = 0, lambda_v = 0, task = "classification")
fm$partial_fit(x, y, ...)
fm$predict(x, ...)
}
}

\section{Methods}{

\describe{
  \item{\code{FM$new(learning_rate = 0.2, rank = 8, lambda_w = 1e-6, lambda_v = 1e-6, task = "classification")}}{Constructor
  for FactorizationMachines model. For description of arguments see \bold{Arguments} section.}
  \item{\code{$partial_fit(x, y, ...)}}{fits/updates model given input matrix \code{x} and target vector \code{y}.
  \code{x} shape = (n_samples, n_features)}
  \item{\code{$predict(x, ...)}}{predicts output \code{x}}
}
}

\section{Arguments}{

\describe{
 \item{fm}{\code{FM} object}
 \item{x}{Input sparse matrix - native format is \code{Matrix::RsparseMatrix}.
 If \code{x} is in different format, model will try to convert it to \code{RsparseMatrix}
 with \code{as(x, "RsparseMatrix")} call}
 \item{learning_rate}{learning rate for AdaGrad SGD}
 \item{rank}{rank of the latent dimension in factorization}
 \item{lambda_w}{regularization parameter for linear terms}
 \item{lambda_v}{regularization parameter for interactions terms}
 \item{n_features}{number of features in model (number of columns in expected model matrix) }
 \item{task}{ \code{"regression"} or \code{"classification"}}
}
}

\keyword{datasets}
