% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-flow.R
\docType{data}
\name{LinearFlow}
\alias{LinearFlow}
\title{Linear Flow method for collaborative filtering}
\format{\code{R6Class} object.}
\usage{
LinearFlow
}
\description{
Creates model which seeks for item similarity matrix
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
  model = LinearFlow$new( rank = 8L,
                          lambda = 0,
                          svd_solver = c("irlba", "randomized_svd"),
                          orthogonal_basis = NULL)
  model$fit_transform(x, ...)
  model$predict(x, k, n_threads = 1L, not_recommend = x, ...)
  model$components
  model$orthogonal_basis
}
}

\keyword{datasets}
