% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{matmult}
\alias{matmult}
\alias{\%*\%,dgRMatrix,matrix-method}
\alias{tcrossprod,dgRMatrix,matrix-method}
\alias{\%*\%,matrix,dgCMatrix-method}
\alias{crossprod,matrix,dgCMatrix-method}
\title{matmult}
\usage{
\S4method{\%*\%}{dgRMatrix,matrix}(x, y)

\S4method{tcrossprod}{dgRMatrix,matrix}(x, y)

\S4method{\%*\%}{matrix,dgCMatrix}(x, y)

\S4method{crossprod}{matrix,dgCMatrix}(x, y)
}
\arguments{
\item{x, y}{Numeric/float matrices.}
}
\value{
A dense \code{matrix}
}
\description{
Multithreaded Sparse-Dense Matrix Multiplication
}
\details{
Accelerates sparse-dense matrix multiplications using openmp. Applicable to
(\code{dgRMatrix}, \code{matrix}), (\code{matrix}, \code{dgRMatrix}),
(\code{dgCMatrix}, \code{matrix}), (\code{matrix}, \code{dgCMatrix}) combinations
}
